style(input): styling custom size and color & remove unnecessary code
