--- src/components/Input/Radio/size.ts
@@ -1,3 +1,0 @@
-type SizeType = 'small' | 'middle' | 'large' | undefined;
-
-export type { SizeType };
